+++ core/auth/__init__.py
@@ -4,5 +4,7 @@ from flask import Blueprint
 
 auth = Blueprint('auth', __name__, url_prefix='/auth')
 
-from . import routes
+from . import utils
 from . import models
+from . import routes
+from . import decorators
